requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
